Fix Update cmd to handle empty i/p data file w/o crashing.
In GetCommand(), if no bookmarks are found in the i/p file
and user asks to delete or update a bookmark,
then print a warning msg & show the menu.

In UpdateCurrentRecord(), handle exceptions better, w/ more
specific warning msg.
Also, do not print a record when the key for the url_dict
is not valid.

Modify some other WARN text msgs.