Add select_recs() func to find records based on keyword.
This function's goal is to select a subset of all the records
in the data set to give a smaller set of data to study and
test.  For example, selecting all records with the keyword
'beginner' might give a set of questions and answers that
can be evaluated manually in a short time.
This will help me to evaluate
my approach and ideas and coding and workflow more quickly,
and develop useful s/w.

Save WIP before stopping work.  Function is partly completed.

Find questions that contain the keyword in Title or Body field.

Need to process Answers also.