First commit for nltk_ex22.py, example using NLP tech.
This program uses the Natural Language Toolkit to perform natural
language processing tasks on the input data.

Uses bag of words code from kaggle.com as a model.

Saving the work in progress now.  It runs and produces output with
no obvious warnings or errors.  It seems to work as expected, but
I have not checked the results to verify that they are correct.

I removed much debugging code and cruft before this commit, and
much remains.  Much clean up is needed.