Change many empty lines to hold a TAB char.
This change allows me to view the file's structure rather easily.

The file appears to have empty lines when viewed in normal mode w/ a
text editor.  When viewed as a VimOutliner file, the content of the file
can be collapsed and expanded easily based on indentation, to show its
structure.  This can be a useful feature for understanding the code and
for debugging.