fga: Save work-in-progress w/ many changes.
Save current code in fga*.py to this
temporary repository (0504_argparse_popular_ques) before making
many changes to it in the work repo.
If I need the code later, it will be in this branch.

This commit has code related to finding popular questions, much
of which should remain in the program.  However, many
lines may be re-arranged or removed due to a design change.
If that change leads to problems, this saved code might be useful.
If the change is OK, this branch can be deleted.