Add new func read_and_grade_answers().
This function shows a question and one answer to the user,
and asks them for a quality grade for the answer: how well
does it answer the question?

Some comments in the program provide a high-level specification
of what the function should do, but there is no code for it yet.

The function is called from main(), only if the cmd line
option '--user_eval' is specified.  After it finishes and
returns control to main(), the program is terminated.