Detect corrupt data when reading ungraded_answers_df.
Add a positive test for an answer record, ie, a record that has
no Title field.

Add a third branch to the logic, to handle the case when data may
be corrupted or when something has gone wrong.  This branch should
never be reached.  Stop the program at that point.