Show menu when user presses '?' key.
Show user some help text, for the '?' key.

Cannot use the 'h' key for help because it marks the current answer
as having high value.  This is too bad, because 'h' is an often-used
convention to show help.  I'll use it for high value because that
fits well with the use of the program.