Accumulate grades for answers each time the program is run.
Do not erase all stored grades when the program is started,
by using an i/p file that has no grades.

Save all the grades & notes that are added when the program runs
in user_eval mode (CLI option -u), to a disk file.   Then read
that file into the df when user_eval mode is next used, so that
grades for answers are kept across runs of the program.

Save df w/ grades to disk before quitting.

Rename save_grade() to store_grade(); & move the code that writes
to disk file from it, into the menu code that handles the 'q'
for quit option.