Add 's' cmd to show current Q&A.
Fix the menu and help cmds to prevent them from advancing the loop
to read the next item from the ungraded dataframe.

Also, prevent them from showing Q & A data: only show that
data when user requests it, w/ the new 's' cmd.  Q & A data are
also shown when program first starts.

Add code after end of if-else clause, to ask for next cmd
from the user & accept it.

Shrink the size of the Q & A shown on screen to the first
55 chars of each.