Copy fga*.py to new program grade_each_answer.py.
First commit of the new program after copying from fga*.
Many lines will be removed.

Plan to move the grading functionality from fga*
to this separate program, to keep fga* focused on analysis tasks.

The grading functionality is more of a utility function
that is done once for a given set of answers, then it is
not needed.  It now clutters the fga* code.  Moving it out
of fga* will make fga* easier to develop and debug.