Clear user_cmd to prevent endless while loop.
The 'else' block of the command interpreter now catches invalid
input and handles it properly.  If user enters an invalid cmd,
this block prints the input it got, then the menu of valid cmds.
Resetting the user_cmd variable stops the main while loop from
continuing infinitely; and allows the secondary while loop to prompt the
user again.

Comment out most of the function calls in main() - these are
probably not needed for the limited functions done in this program.
Preliminary tests show that the code works OK.