grade_each_answer: Add cmd line option '-d' for debugging.
Add '-d' to enable debugging, which only affects one new variable.

When running normally, users will want to see the entire question
and answer, to grade it.  In this case, do not specify '-d'.

When debugging, developers might not want to see the full Q&A,
which can be dozens or hundreds of lines long.

Set debug mode to limit the size of the Q&A that are shown on
screen.  The value of last_char is now '55', so the first
55 chars of each Q&A are shown on screen.