Set last_char to a specific size for non-debug mode.
When not in debug mode, the max size for question & answer
text shown on screen is set to 999,999.  This should be
large enough to accommodate most if not all Q&A of interest.

Detect if length of Q or A, when operating normally (ie, not
in debug mode) exceeds the limit.  If so, issue a warning.

If in debug mode, disable those warnings.