Save graded dataframe before exit, if quit cmd is not used.
When all answers have been graded, the program terminates.
Previous code was broken, and any changes made during that
last run were not saved, because the quit cmd was not executed,
and that was the only way to save the o/p.

The code now saves data when read_and_grade_answers() is finished,
just before it returns control to its caller.