grade*: Add grade 'e'; change 'b' to 'very good'; update header notes.
Change the grading system from five levels to six levels,
to distinguish very good, good, and fair answers better.

The quality of an answer depends on content, writing, organization,
and presentation.  These six levels will allow, eg, an answer
with helpful content but poor writing and organization (ie, it
may be difficult to understand) to be marked different from an
answer with poor content that is more readable.

Edit some comments in the program's header to show the current
content of the o/p file's Q & A.