nltk: Use global vars from config.py file.
Update nltk_ex25.py with changes that were proven in the experiments
with n25_nltk.py.  Most of the changes relate to using global vars
for the project, now specified in the config.py file.

The n25_nltk.py file will not be committed, and will be deleted.
The nltk_ex25.py program will be maintained from now on, and
kept current in the git repo.