Return a df of answers with hi-score terms from a func.
The search_for_terms() func now returns a df containing selected
answers, for further review.  This is a step towards fixing the
bug that only saves the last group of the many Q&A groups that
the program can now process.

Also, change some variables to increase the number of answers
to be checked for good quality data.