Fix bug: concatenate all df's of Q&A groups into one df & o/p file.
Prior s/w only saved the df of the last Q&A group to be processed.
That bug is fixed, and all such groups are saved into a single file.
That file may not be needed, after further work to process the
data uses the df that is already available.