fga: Remove mention of 'run' cmd.
Remove the 'run' cmd from instructions about debugging problems.
It is not available.

Include the ' pd.set_option('display.max_colwidth', 100)'
statement to modify format of some output data on screen;
but comment it out in fga.  Keep it active in the grade*.py
program, based on experiments with it.