Plot a scatter matrix of all features against each other.
Use the pandas plotting function to make the scatter matrix.
Use this matrix for a quick but limited overview of the data.

Add menu option 'dm' to request the scatter matrix plot.

Add a function draw_scatter_matrix_plot().

Modify the function draw_scatter_plot() to include four arguments,
to make it more general, for future plotting.

Add a function save_prior_file(), to make a backup copy
of a file with the '.bak' extension.

Add menu option 'dr' to request a plot of reputation data.
The code for this plot is not ready.