Replace global cf.all_ans_df w/ local qagroup_df.
This fixes a temporary change that assigned cf.all_ans_df to qag_df.

Most of the needed changes are in the nltk module.

Comment out the main() func in the nltk module.
The nltk module provides NLP functionality to its callers,
and it need not run as a stand-alone program, so this and
other support code is being removed.

The code works OK.  Some cleanup will be done next.