Update dirs & files for static web site at Github Pages.
Remove the README files that were copied from github project dir.
If they are needed in the web site, link to them.

Remove the _config.yml file from the learn_python/ dir,
and rely on that file in the docs/ subdir only.

Changed the index.md file (Home page) to be smaller.  Its main
purpose now is to point to the Find Answers project.