Calculate owner reputation.
Add several functions to calculate the reputation of each OwnerUserId,
based on the scores given to their answers, and to organize
and display some other statistics:
build_stats(), calculate_owner_reputation(), and gd2_group_data().

Add two new menu entries, cr and cs for calculate reputation
and calculate statistics, to calculate and to display
statistical data in a table and on a scatter matrix plot.