First commit of a file w/ table of numeric data.
new file:   docs/find_answers/qa_stats.html

This HTML table does not contain the Title, Body, or Hi Score Terms
data.  It is a compact representation of some statistics about
the Q&A groups in the data set.  It can be used to compare
Scores and Owner Reputation and hstCounts.