Increase slice sizes to build pop_and_top_l.
That change is needed when using the bigger data files
(a6 and q6 instead of a3 and q3).  It was made in
select_questions(), to increase the size of the
popular-questions-and-top-owners list, to get more
o/p data.

With the slice sizes of 900 and 900, the output included data
from 23 questions and 223 answers.

Make all_ans_df an argument for show_menu(), to use in
calculate_reputation(), so the code does not need to re-read
the i/p answers file a second time.

Write more data to the log file, for easier debugging, and
to have a historical record of changes and their effects.