Add cmd line option to quit the program.
Specifying the quit option lets the program start and run and write the
first set of output files to disk; then it stops before it shows the
user menu on the console.  This behavior allows a simple automated test
to run to completion, without having the user menu interfere with its
operation.