Rename own_rep_file to stats_fname.
Use a better name for the variable name used when writing
to two o/p files.  The original name is used elsewhere
in the program, and it is confusing to use it here also.

Also, round() a calculated value so there is less chance
of a failed diff operation (as used by test tools when
comparing data to determine pass or fail).  The rounded
numbers also look better on the screen.