Define clean_q_a earlier, to eliminate a function call.
In the find_freq_words() function:
Move the definition of clean_q_a up several lines and use it
in two places, so that a second call to convert_text_to_words()
is not needed.

Rename var clean_body to clean_body_s, to indicate that it
is a string.

Update some docstrings and other comments for clarity and
accuracy.