fga: Update docstring for file & for main().
Add details about the actions of the program, to make it easier
to understand the code and its purpose and how it should work.
This data will also help in future refactoring work being planned.

Add a few notes on requirements to the header.

Clean up some spacing and indent problems.