Rename sort_answers_by_score() to sort_q_a_by_score().
Change func name to reflect its current action: it sorts
both the question and its answers in a Q&A group by their
scores..

Add comment that this is a temporary change, and in the future
the code should be modified to use only answers and not include
questions.  See b.98.