fga: Save records w/ keyword to disk files.
Enhance the search functionality to save full records to disk.
A full record is a question or answer that includes the Body text and
most other fields.  Data is saved to a csv file and an html file.
Those files are overwritten with each successful search operation,
so the user must save them manually if required.

Rename variable row_short_t to row_for_summary_t, to clarify its
use.  It is a tuple that contains six fields of a record, and does not
include the Body text.