fga: work-in-progress, basic search function is OK.
wip: Save this code in its current state before doing major cleanup,
because it works.

Many changes made to the code to provide the search tool and to build
an html o/p file w/ the discovered data in the specified order.
The basic search function now works.

This commit includes much cruft, obsolete comments and code.

The func select_keyword_recs() contains code for the search and display
functionalities.  It has some functionality that might be implemented
in other functions in this file.