fga: Call select_keyword_recs() from sga*py instead of local func.
These changes allow fga to call the search function in the
sga*py module, so it can be removed from fga*py.

Import sga_show_good_answers as a module into fga, to have access
to the search func.

Define 'args' in fga*py to hold parser data (copy code from sga*py,
where it was debugged).

Add 'args' to signature of select_keyword_recs() and to the code
that calls that search function.