Replace the variables args and cli_args_d with opt_ns.
Change both fga and sga to use this new name.

This change avoids a conflict with the 'args' function in the pdb
debgger (see git commit 2017-10-15 13:56, e66cc8 for details).

Use the new name to replace hard-coded values that were used
for debugging.

fga: Add the command-line search option to get_parser().