First commit of search o/p file; for keyword=function.
Add new file to GitHub Pages site for the project.

The fga program's search facility was enhanced recently, and this
file is an example of the o/p that it produces.

It is the result of searching for the word 'function', in the
small data files used for development work (q3_992.csv and
a3_986.csv).