Edit Gemfile.lock to specify nokogiri v1.8.1.
Modify this file that the project uses for the GitHub Pages site.
It probably came from the Scribble repository, which is a theme
for Jekyll, which is the tool to build the static web pages for GH Pages.

Change from older versions because of a security alert from github:

    Known critical severity security vulnerability detected in nokogiri
    < 1.8.1 defined in Gemfile.lock.
    Gemfile.lock update suggested: nokogiri ~> 1.8.1.

See the github repository at muan/scribble/Gemfile.lock; it
does not specify nokogiri v1.8.1, so I hope this change is complete
and correct.

My system now has v1.8.1 installed and the local testing shows
that the site works OK.