sga: Print path to file w/ search o/p data to screen.
After a search request has completed, tell the user which file
to open to see the result of the latest search request.
The program shows a 1-line per record summary of the search
response on screen, but the full o/p is only available
in a file that the user must view manually.

Also, delete two unrelated and unneeded lines from fga*py.