sga: Write html o/p file w/ all df columns, like csv file.
Write data to a new file, search_result_all_fields.html.

Having the two different files w/ the same columns lets us compare
them and expect that the content matches.

A simple test that counts occurrences of the search term in the csv and
html files was failing, because the html file did not have all the
columns of the csv file.  The simple test now passes.