Modify order of df columns shown in some o/p stmts.
Show columns in a standard order in most cases, that indicates the
sort order of the data: column 1 is first key, col 2 is second, and
col 3 is third, which are usually the HSTCount, Score, Id columns.

Move some print clauses and add a title at top of the printed output.

Change one if-clause from using the 'veryverbose' option,
to using the 'debug' option.

Most of these changes are to make debugging easier by showing
useful data on screen when using the CLI verbose or debug options.