config: Move some global vars into this config file.
Move the constants out of fga*py and write.py to config.py.

In the various modules that use these vars, assign the vars
to the global value defined in config.py.

If a change to a global var is needed, it can be made in the config.py
file, and will be retrieved from there by other modules that import
it.