fga: Move keyword checking from main() to a new func.
Create the check_for_keyword() function, and move about a dozen
lines of code out of main() and into it.

Add a condition to the check, to detect that the o/p dataframe
has the value 'None', when the keyword is not found.  This prevents
an error msg from being shown to the user in some cases.

De-clutter main().