Fix return values; rename progress_msg_factor to progress_i.
sga: Return dataframes from select_keyword_recs() and remove check code
that is not needed.
Returning 'None' to the caller caused problems; this prevents that
problem.

In several modules and funcs:
Rename progress_msg_factor to progress_i.

Use that var in several places.
Fix a problem of calculating two different values for it, and using
the wrong one.  Now progress msgs are updated on screen for most
if not all cases.

Current code does not show a progress msg while reading i/p file, but
some code is left in place for that.

The progress_i value is declared global in main(), and is also used
in the argument lists of several functions.
TBD: Refactor that code.  This code now works for the tests I did.

Add log messages to indicate progress; in some places, these statements
replace comments, which are redundant.