sga: Rename a field from RecId to LineNum.
Rename RecId to prevent confusing it with the variable name
'rec_id' (used for the Id value),
which is used in the code and is unrelated to the line number.

LineNum is a more accurate name for this column.  The name is
not now used in the code; it is only used in a comment
to identify that field.