Write only the Id field to a file used for testing.
Problem: a test was failing because the two files it compared
had the same data, but not in the same order.  The data files
contained a dataframe with six columns.

Soln:
.  Remove the problem file from the ref/ subdir, qa_with_keyword.csv.
.  Write new o/p file to disk that has only the Id field
for each desired record; save a good copy of that file in the ref/ dir
where pytest will find it.
.  Write this o/p file from two places in the code (in fga*py
and in text_ui.py).