Write test o/p sorted by Id value to disk file.
Fix bug where the actual and reference o/p had the same content,
but not in the same order.

Only save the Id value of the question and answer records, and sort that
data to make them easy to compare.

Change two files that have the code to save this data to disk.

Keep the old code that saves six columns of the dataframe in case
it is needed; but comment it out.