fga: Simplify read_data(); rename to read_df_in_csv_file().
Read only one file when this function is called.

Make the func simpler and smaller and more general, and name
it more specifically.

Remove unneeded code.

text_ui: Fix bug that crashed the program when search term
is not found in the data.

Detect that the returned df is empty, show error msg,
and prompt for a new search term.