Remove global progress_i; calculate it only where used.
This var is used to show progress messages, to show the user
that the program is running and not hung.

Do not use one value for all such msgs; re-calculate it based
on the size of the data being processed at the point of use.

This change affects three files, and simplifies args for
several functions, and removes one global var.

Some progress msgs are shown when the fga program is invoked
with the '-v' for verbose option.