Write o/p file w/ all data in one column.
Include each Q&A group in the same order that is shown
on the stackoverflow.com web site: q.title, q.body, a1, a2, ...

Include all Q&A groups in the i/p dataframe.

Make a list w/ the data and convert it to a df, then use
the existing write_df_to_html() function.