fga: Remove numlines from the read_df*() func.
Do not return numlines from the read_df_in_csv_file() function,
it is not needed by the calling code.

Remove it from the calling statements for read_df_in_csv_file().

Remove that var from the read*() function, and replace it in the
one place that it is used with a call to the len() function.