WIP: commit working and ugly code before major chgs.
Work In Progress commit.

The code has some partial improvements for the outline-formatted
file, the *.otl file.  Eg, the question body text is indented below a level
called 'Q.Body', which allows a less-cluttered and more-focused
display of text.  Make a similar change to each answer.

There are many lines of cruft to remove or change.

Some working code will be refactored, and this commit saves it if
needed later.