search_for_keyword: Return empty df instead of None object.
Old code found an empty dataframe, but simply returned to caller a
None object.  This forces the calling code to detect
a None object, for which I used a try/except block (in 4 places).

Return the empty df instead of None, so that the existing test for
empty df is adequate in the calling code.

save_output.py, text_ui.py:
Comment out the try/except blocks in these files.