sga: Modify the parameter list of select_keyword_recs().
Modify the order of vars in that parameter list.
Change this list in sga, and in the one place that calls it,
sfk.search_for_keyword().

Make this change so that the sequence of the variables in
the list is the same as the parameter list of the
search_for_keyword() func.

Do this to prepare to replace calls to sfk.search_for_keyword()
with calls to sga.select_keyword_recs().