misc.py: Move calc_progress() from config.py to misc.py.
Add new file to git repo: util/misc.py.
Move one func into it.

Modify imports and function calls in 3 programs that use
calc_progress(), to get it from its new location.

Use util/misc.py to hold utility programs that are used elsewhere
and are not big or special enough to be in a separate file.