misc.py: Move strip_html() from nltk_ex25.py to util/misc.py.
Make the strip program easily accessible from a utility file,
instead of from a main file of the fga project.  This utility
function is used elsewhere and should not be in a main program
of the fga project.

Change the imports and calls to strip_html() in 2 programs,
to use it from its new location:
nltk_ex25 and util/write.