First commit of check_urls code & data.
A simple program to find http urls in a file (asciidoc format),
then query each one and print the result.  Use it to find dead
links.

readme.asciidoc is a sample input file with some links.  Copy it
to /tmp/ to use it with this program, or edit the code to use
a different path/infile.