Rename file from readme.asciidoc to test_data.asciidoc
Update the input file name in check_urls to use the new name.

Rename the file because github shows the content of a
'readme' file on screen whenever it finds one in a project
page.  In this case, the readme file contains test data, and
not text about the project.